+++ engine/src/test/kotlin/dev/m1sk9/lunaticChat/engine/settings/SettingsDataClassesTest.kt
@@ -1,6 +1,5 @@
 package dev.m1sk9.lunaticChat.engine.settings
 
-import dev.m1sk9.lunaticChat.engine.converter.CacheData
 import kotlinx.serialization.json.Json
 import java.util.UUID
 import kotlin.test.Test
@@ -103,29 +102,5 @@ class SettingsDataClassesTest {
         val data = PlayerSettingsData(japaneseConversion = mapOf(testUUID to true))
         val serialized = json.encodeToString(PlayerSettingsData.serializer(), data)
         assertTrue(serialized.contains(testUUID.toString()))
-    }
-
-    // --- CacheData ---
-
-    @Test
-    fun `CacheData should store version and entries`() {
-        val data = CacheData(version = "1.0", entries = mapOf("hello" to "こんにちは"))
-        assertEquals("1.0", data.version)
-        assertEquals("こんにちは", data.entries["hello"])
-    }
-
-    @Test
-    fun `CacheData serialization round-trip should preserve all fields`() {
-        val original =
-            CacheData(
-                version = "2.0",
-                entries = mapOf("hello" to "こんにちは", "world" to "世界"),
-            )
-
-        val serialized = json.encodeToString(CacheData.serializer(), original)
-        val deserialized = json.decodeFromString(CacheData.serializer(), serialized)
-
-        assertEquals(original.version, deserialized.version)
-        assertEquals(original.entries, deserialized.entries)
     }
 }
