+++ platform-paper/build.gradle.kts
@@ -20,6 +20,9 @@ dependencies {
 
     // Paper-specific dependencies
     compileOnly("io.papermc.paper:paper-api:26.2.build.92-stable")
+    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.11.0")
+    implementation("io.ktor:ktor-client-core:3.5.2") // Google IME client, update check
+    implementation("io.ktor:ktor-client-cio:3.5.2")
     implementation("com.charleskorn.kaml:kaml:0.104.0") // YAML configuration
     implementation("org.jetbrains.kotlin:kotlin-reflect:2.4.10") // Annotation processing
