+++ platform-paper/src/main/kotlin/dev/m1sk9/lunaticChat/paper/converter/CacheData.kt
@@ -1,4 +1,4 @@
-package dev.m1sk9.lunaticChat.engine.converter
+package dev.m1sk9.lunaticChat.paper.converter
 
 import kotlinx.serialization.Serializable
