+++ platform-paper/src/main/kotlin/dev/m1sk9/lunaticChat/paper/converter/KanaConverter.kt
@@ -1,4 +1,4 @@
-package dev.m1sk9.lunaticChat.engine.converter
+package dev.m1sk9.lunaticChat.paper.converter
 
 /**
  * Converts romanji text to hiragana using Trie data structure.
